+++ tests/Predis/Command/Redis/VEMB_Test.php
@@ -66,6 +66,7 @@ class VEMB_Test extends PredisCommandTestCase
     /**
      * @dataProvider quantisationProvider
      * @group connected
+     * @group relay-incompatible
      * @return void
      * @requiresRedisVersion >= 8.0.0
      */
@@ -98,6 +99,7 @@ class VEMB_Test extends PredisCommandTestCase
     /**
      * @dataProvider quantisationProvider
      * @group connected
+     * @group relay-incompatible
      * @return void
      * @requiresRedisVersion >= 8.0.0
      */